sam2-main/requirements.txt
# 基础核心包
torch==2.9.1+cu126
torchvision==0.24.1+cu126
timm==1.0.24

# sam关键模块
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==12.0.0
matplotlib==3.10.8
pyarrow
pandas